backend/requirements.txt
# Core Framework
fastapi==0.121.0
uvicorn==0.38.0
pydantic==2.12.4

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# File handling
python-multipart==0.0.9

# HTTP and networking
httpx==0.27.0

# AI/ML
google-generativeai==0.8.5

# Environment variables
python-dotenv==1.0.0

# Database
supabase==2.10.0

# Existing dependencies
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
h11==0.16.0
idna==3.11
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==24.3.0
Authlib==1.6.5
Automat==24.8.1
babel==2.17.0
beartype==0.22.5
beautifulsoup4==4.12.3
bleach==6.2.0
bs4==0.0.2
cachetools==6.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
comm==0.2.2
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
cyclopts==4.2.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22.3
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fake-useragent==2.0.3
fastapi==0.115.11
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastmcp==2.13.0.2
filelock==3.16.1
fonttools==4.58.5
fqdn==1.5.1
google-generativeai==0.8.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
incremental==24.7.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itemadapter==0.10.0
itemloaders==1.3.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.5
kagglehub==0.3.12
keyring==25.6.0
kiwisolver==1.4.8
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mcp==1.21.0
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.8.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.0
notebook_shim==0.2.4
numpy==2.3.1
openai==2.7.1
openapi-pydantic==0.5.1
opendatasets==0.1.22
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parse==1.20.2
parsel==1.9.1
parso==0.8.4
pathable==0.4.4
pathvalidate==3.3.1
pdfminer.six==20251107
pdfplumber==0.11.8
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.51
Protego==0.3.1
protobuf==6.31.1
proxyscrape==0.3.0
psutil==7.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyDispatcher==2.0.7
pyee==11.1.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.3
pypdfium2==5.0.0
pyperclip==1.11.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.7.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-rst==1.3.2
rich-toolkit==0.13.2
rpds-py==0.24.0
Scrapy==2.12.0
scrapy-rotating-proxies==0.6.2
Send2Trash==1.8.3
service-identity==24.2.0
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.46.1
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
tldextract==5.1.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.11.0
typer==0.20.0
types-python-dateutil==2.9.0.20241206
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
uv==0.9.8
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.2.1
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
zipp==3.21.0
zope.interface==7.2
